Removed cron_wrappers.py
